made the retrieved length an int instead of a long

+++ jacs2-services/src/main/java/org/janelia/jacs2/dataservice/storage/StorageService.java
@@ -230,7 +230,7 @@ public class StorageService {
                                                      String authToken,
                                                      int level,
                                                      long offset,
-                                                     long length) {
+                                                     int length) {
         Client httpclient = HttpUtils.createHttpClient();
         try {
             WebTarget target = httpclient.target(storageURI).path("list");
